### 2025-04-17 #238 @ManonMarchand
MOCpy
Submitting Author: (@ManonMarchand)
All current maintainers: (@fxpineau, @tboch, @bmatthieu3, @ManonMarchand)
Package Name: MOCPy
One-Line Description of Package: MOCPy allows to read/write, and create Multi-Order Coverage Maps
Repository Link:  https://github.com/cds-astro/mocpy
Version submitted: 0.17.1
EiC: (@coatless, @eliotwrobson)
Version accepted: TBD

---

MOCPy is a python package that allows the manipulation of Time-Space Multi-Order Coverage maps. These maps are defined by an International Virtual Observatory standard (https://ivoa.net/documents/MOC/20220727/index.html).

Concretely, time and space (on the sky sphere) are described by cells of different coarseness, and a MOC is an ensemble of these cells.

Here are three MOCs (plotted with MOCPy) describing regions of the sky :

This format is particularly useful to perform operations (union, difference, ...) on time and space regions.

Here is the union of the previous MOCs:

 Typical uses of the library are:

- estimation of telescopes field of view (uses space information)
- knowing if a satellite has observed a short-lived event in the sky (uses time and space information)

We were an astropy-affiliated package with the former system (pre-2024).

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The audience is astronomers. From a quick scanning of the libraries that use MOCPy on github, we identify three major poles:

 1. time-dependent astronomy fields (gravitational waves, transients, multi-messenger...).
 2. big astronomical databases. They either cross-match astronomical catalogs between each other, or they are building and interrogating databases indexed with the HEALPix tesselation (the one used for MOC space cells).
 3. telescope crew that use MOCPy to calculate field of view and plan observations

There is pymoc (https://github.com/grahambell/pymoc), a pure python implementation. It works, but the API is more minimal (no creation of MOCs from shapes, no check to see if a sky-coordinate falls within a MOC, only Space-MOCs no Time-MOCs, no plotting utilities, no multi-threading, ...)
