### 2025-05-12 bec3938
maint: replace deprecated abstractclassmethod and abstractclassproperty decorators
