### 2025-05-12 c0cce87
maint: fix version in pyproject to build with maturin 1.8
see https://github.com/PyO3/maturin/issues/2416
