### 2025-05-13 aa10b3f
maint:add cargo fmt hook
