+++ src/sfmoc.rs
@@ -1,0 +1,228 @@
+//! Methods releted to SF-MOC
+
+use numpy::{PyArrayMethods, PyReadonlyArrayDyn};
+use pyo3::{
+  exceptions::{PyIOError, PyValueError},
+  pyfunction, PyResult,
+};
+
+use moc::storage::u64idx::U64MocStore;
+
+/// Deserialize a Frequency-Space coverage from a FITS file.
+///
+/// # Arguments
+///
+/// * ``index`` - the index used to store the Frequency-Space coverage
+/// * ``path`` - the FITS file path
+///
+/// # Warning
+///
+/// This function is not compatible with pre-v2.0 MOC standard.
+///
+/// # Errors
+///
+/// This method returns a `PyIOError` if the the function fails in writing the FITS file.
+#[pyfunction]
+pub fn coverage_sf_from_fits_file(path: String) -> PyResult<usize> {
+  U64MocStore::get_global_store()
+    .load_sfmoc_from_fits_file(path)
+    .map_err(PyIOError::new_err)
+}
+
+#[pyfunction]
+pub fn sfmoc_from_fits_raw_bytes(raw_bytes: &[u8]) -> PyResult<usize> {
+  U64MocStore::get_global_store()
+    .load_sfmoc_from_fits_buff(raw_bytes)
+    .map_err(PyIOError::new_err)
+}
+
+/// Deserialize a Frequency-Space coverage from an ASCII file (compatible with the MOC v2.0 standard).
+///
+/// # Arguments
+///
+/// * ``index`` - the index used to store the Frequency-Space coverage
+/// * ``path`` - the ASCII file path
+///
+/// # Errors
+///
+/// This method returns a `PyIOError` if the the function fails in writing the FITS file.
+#[pyfunction]
+pub fn coverage_sf_from_ascii_file(path: String) -> PyResult<usize> {
+  U64MocStore::get_global_store()
+    .load_sfmoc_from_ascii_file(path)
+    .map_err(PyIOError::new_err)
+}
+
+/// Deserialize a Frequency-Space coverage from an JSON file.
+///
+/// # Arguments
+///
+/// * ``index`` - the index used to store the Frequency-Space coverage
+/// * ``path`` - the JSON file path
+///
+/// # Errors
+///
+/// This method returns a `PyIOError` if the the function fails in writing the FITS file.
+#[pyfunction]
+pub fn coverage_sf_from_json_file(path: String) -> PyResult<usize> {
+  U64MocStore::get_global_store()
+    .load_sfmoc_from_json_file(path)
+    .map_err(PyIOError::new_err)
+}
+
+/// Deserialize a Frequency-Space coverage from an ASCII string (compatible with the MOC v2.0 standard).
+///
+/// # Arguments
+///
+/// * ``index`` - the index used to store the Frequency-Space coverage
+/// * ``ascii`` - the ASCII string
+///
+/// # Errors
+///
+/// This method returns a `PyIOError` if the the function fails in writing the FITS file.
+#[pyfunction]
+pub fn coverage_sf_from_ascii_str(ascii: String) -> PyResult<usize> {
+  U64MocStore::get_global_store()
+    .load_sfmoc_from_ascii(&ascii)
+    .map_err(PyIOError::new_err)
+}
+
+/// Deserialize a Frequency-Space coverage from an JSON string.
+///
+/// # Arguments
+///
+/// * ``index`` - the index used to store the Frequency-Space coverage
+/// * ``json`` - the JSON string
+///
+/// # Errors
[…]
