### 2025-05-13 0f1ad6c
fix: use one-liners in pytest.raises
