### 2025-05-13 d4781a3
feat: add refine to order method for all 1D MOCs
