### 2025-05-13 7913d82
feat: add refine to order method for all 1D MOCs
