### 2025-05-13 f4236e9
fix: update image in linux CI
