### 2025-05-14 8a5fb32
fix: pin rust version for emscripten build
