+++ .github/workflows/emscripten.yml
@@ -15,23 +15,24 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v5
         with:
-          python-version: '3.10'
+          python-version: '3.12'
       - run: pip install pyodide-build
       - shell: bash
         run: |
          echo EMSCRIPTEN_VERSION=$(pyodide config get emscripten_version) >> $GITHUB_ENV
          echo PYTHON_VERSION=$(pyodide config get python_version | cut -d "." -f 1-2) >> $GITHUB_ENV
-      - uses: mymindstorm/setup-emsdk@v12
+      - uses: mymindstorm/setup-emsdk@v14
         with:
           version: ${{ env.EMSCRIPTEN_VERSION }}
           actions-cache-folder: emsdk-cache
-      - name: Build wheels
-        uses: PyO3/maturin-action@v1
-        with:
-          target: wasm32-unknown-emscripten
-          args: --release --out dist -i ${{ env.PYTHON_VERSION }}
-          sccache: 'true'
-          rust-toolchain: nightly
+      - shell: bash
+        # rust version is pinned for compatibility with emscripten 3.1.58
+        run: |
+          pip install maturin
+          rustup install nightly-2025-02-01
+          rustup default nightly-2025-02-01
+          rustup target add wasm32-unknown-emscripten
+          maturin build --target wasm32-unknown-emscripten -i 3.12
       - name: Upload wheels
         uses: actions/upload-artifact@v4
         with:
