### 2025-05-15 a119e85
fix: some stmocs rust methods changed names due to the introduction of sfmocs
