+++ python/mocpy/tmoc/tmoc.py
@@ -18,16 +18,19 @@ DAY_MICRO_SEC = 86400000000.0
 
 def times_to_microseconds(times):
     """
-    Convert a `astropy.time.Time` into an array of integer microseconds since JD=0, keeping the microsecond resolution required for `~mocpy.tmoc.TimeMOC`.
+    Convert a `astropy.time.Time` into an array of integer microseconds since JD=0.
+
+    This keeps the microsecond resolution required for `~mocpy.TimeMOC`.
 
     Parameters
     ----------
     times : `astropy.time.Time`
-    Astropy observation times
+        Astropy observation times
 
     Returns
     -------
-    times_microseconds : `np.array`
+    `np.array`
+        Time in microseconds
     """
     times_jd = np.asarray(times.jd, dtype=np.uint64)
     times_us = np.asarray(
@@ -48,7 +51,7 @@ def microseconds_to_times(times_microseconds):
 
     Returns
     -------
-    times : `astropy.time.Time`
+    `astropy.time.Time`
     """
     jd1 = np.asarray(times_microseconds // DAY_MICRO_SEC, dtype=np.float64)
     jd2 = np.asarray(
@@ -134,7 +137,7 @@ class TimeMOC(AbstractMOC):
 
         Returns
         -------
-        moc : `~mocpy.tmoc.TimeMOC`
+        `~mocpy.TimeMOC`
             The degraded MOC.
         """
         index = mocpy.degrade(self.store_index, new_order)
@@ -186,8 +189,7 @@ class TimeMOC(AbstractMOC):
 
         Returns
         -------
-        moc : `~mocpy.tmoc.TimeMOC`
-            The MOC
+        `~mocpy.TimeMOC`
         """
         index = mocpy.new_empty_tmoc(np.uint8(max_depth))
         return cls(index)
@@ -205,8 +207,7 @@ class TimeMOC(AbstractMOC):
 
         Returns
         -------
-        moc : `~mocpy.tmoc.TimeMOC`
-            The MOC
+        `~mocpy.TimeMOC`
         """
         ranges = np.zeros((0, 2), dtype=np.uint64) if ranges is None else ranges
 
@@ -237,7 +238,7 @@ class TimeMOC(AbstractMOC):
 
         Returns
         -------
-        time_moc : `~mocpy.tmoc.TimeMOC`
+        `~mocpy.TimeMOC`
         """
         times = times_to_microseconds(times)
         times = np.atleast_1d(times)
@@ -264,9 +265,9 @@ class TimeMOC(AbstractMOC):
 
         Returns
         -------
-        time_moc : `~mocpy.tmoc.TimeMOC`
+        `~mocpy.TimeMOC`
         """
-        # degrade the TimeMoc to the order computed from ``delta_t``
+        # degrade the TimeMOC to the order computed from ``delta_t``
         depth = TimeMOC.time_resolution_to_order(delta_t)
 
         min_times = times_to_microseconds(min_times)
@@ -312,9 +313,9 @@ class TimeMOC(AbstractMOC):
 
         Returns
         -------
-        time_moc : `~mocpy.tmoc.TimeMOC`
+        `~mocpy.TimeMOC`
         """
-        # degrade the TimeMoc to the order computed from ``delta_t``
+        # degrade the TimeMOC to the order computed from ``delta_t``
         depth = TimeMOC.time_resolution_to_order(delta_t)
 
         min_times = np.asarray(min_times.jd)
@@ -357,21 +358,21 @@ class TimeMOC(AbstractMOC):
 
         Parameters
         ----------
-        another_moc : `~mocpy.tmoc.TimeMoc`
+        another_moc : `~mocpy.TimeMOC`
         order_op : int
[…]
