+++ python/mocpy/moc/moc.py
@@ -146,7 +146,7 @@ class MOC(AbstractMOC):
     A MOC corresponds to a list of `HEALPix <https://healpix.sourceforge.io/>`__ cells at different depths.
     This class gives you the possibility to:
 
-    1. Define `~mocpy.moc.MOC` objects:
+    1. Define `~mocpy.MOC` objects:
 
     - From a FITS file that stores HEALPix cells (see `load(path, 'fits')`).
     - Directly from a list of HEALPix cells expressed either as a numpy structural array (see `from_healpix_cells`) or a simple
@@ -155,7 +155,7 @@ class MOC(AbstractMOC):
     - From a convex/concave polygon (see `from_polygon`).
     - From a cone (will be implemented in a next version).
 
-    2. Perform fast logical operations between `~mocpy.moc.MOC` objects:
+    2. Perform fast logical operations between `~mocpy.MOC` objects:
 
     - The `intersection`
     - The `union`
@@ -163,14 +163,14 @@ class MOC(AbstractMOC):
     - The `complement`
 
 
-    3. Plot the `~mocpy.moc.MOC` objects:
+    3. Plot the `~mocpy.MOC` objects:
 
     - Draw the MOC with its HEALPix cells (see `fill`)
     - Draw the perimeter of a MOC (see `border`)
 
-    4. Get the sky coordinates defining the border(s) of `~mocpy.moc.MOC` objects (see `get_boundaries`).
+    4. Get the sky coordinates defining the border(s) of `~mocpy.MOC` objects (see `get_boundaries`).
 
-    5. Serialize `~mocpy.moc.MOC` objects to `astropy.io.fits.HDUList` or JSON dictionary and save it to a file.
+    5. Serialize `~mocpy.MOC` objects to `astropy.io.fits.HDUList` or JSON dictionary and save it to a file.
     """
 
     # Maximum order (or depth) of a MOC
@@ -227,6 +227,10 @@ class MOC(AbstractMOC):
         include_indirect_neighbours : bool
             if `false`, only consider  cells having a common edge as been part of a same MOC
             if `true`, also consider cells having a common vertex as been part of the same MOC
+
+        Returns
+        -------
+        int
         """
         return mocpy.split_count(self.store_index, include_indirect_neighbours)
 
@@ -240,9 +244,13 @@ class MOC(AbstractMOC):
             if `false`, only consider  cells having a common edge as been part of a same MOC
             if `true`, also consider cells having a common vertex as been part of the same MOC
 
-        Warning
+        Returns
         -------
-        Please use `~mocpy.moc.MOC.split_count` first to ensure the number is not too high
+        `~mocpy.MOC`
+
+        Notes
+        -----
+        Use `~mocpy.moc.MOC.split_count` first to ensure the number is not too high
         """
         indices = mocpy.split(self.store_index, include_indirect_neighbours)
         return [MOC(index) for index in indices]
@@ -261,7 +269,7 @@ class MOC(AbstractMOC):
 
         Returns
         -------
-        moc : `~mocpy.moc.MOC`
+        `~mocpy.MOC`
             The degraded MOC.
         """
         index = mocpy.degrade(self.store_index, new_order)
@@ -315,7 +323,7 @@ class MOC(AbstractMOC):
 
         Returns
         -------
-        array : `~np.ndarray`
+        `~np.ndarray`
             A mask boolean array
 
         See Also
@@ -387,7 +395,7 @@ class MOC(AbstractMOC):
 
         Returns
         -------
-        array : `~np.ndarray`
+        `~np.ndarray`
             A mask boolean array
 
         Examples
@@ -522,7 +530,7 @@ class MOC(AbstractMOC):
 
         Returns
         -------
[…]
