+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 repos:
   # General hooks
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.6.0
+    rev: v5.0.0
     hooks:
       - id: detect-private-key
         description: "Check accidental private keys in code"
@@ -13,7 +13,7 @@ repos:
         description: "Remove trailing whitespaces"
   # Fast linter to replace flake8
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: 'v0.6.4'
+    rev: 'v0.11.9'
     hooks:
      # linter
      - id: ruff
