+++ python/mocpy/moc/moc.py
@@ -302,8 +302,7 @@ class MOC(AbstractMOC):
         """
         if new_order <= self.max_order:
             warnings.warn(
-                "'new_order' is less precise than the current max order. "
-                "Nothing done.",
+                "'new_order' is less precise than the current max order. Nothing done.",
                 stacklevel=2,
             )
         mocpy.refine(self.store_index, new_order)
