+++ python/mocpy/fmoc/fmoc.py
@@ -169,8 +169,7 @@ class FrequencyMOC(AbstractMOC):
         """
         if new_order <= self.max_order:
             warnings.warn(
-                "'new_order' is less precise than the current max order. "
-                "Nothing done.",
+                "'new_order' is less precise than the current max order. Nothing done.",
                 stacklevel=2,
             )
         mocpy.refine(self.store_index, new_order)
