### 2025-05-15 0f7c33c
maint: add hook to clean notebooks metadata
