+++ pyproject.toml
@@ -119,7 +119,7 @@ extend-ignore = ["D203", "D213", "D100", "N816", "D105", "N806", "N803", "FBT002
 # D104: Missing docstring in public package
 # D103: Missing docstring in public function
 # D100: Missing docstring in public module
-# B008: Do not perforn function call in argument defaults
+# B008: Do not perform function call in argument defaults
 "__init__.py" = ["D104"]
 "python/mocpy/moc/plot/wcs.py" = ["B008"]
 "python/mocpy/tests/*.py" = ["S101", "D103"] # S101: use of "assert" detected
@@ -132,5 +132,5 @@ convention = "numpy"  # Accepts: "google", "numpy", or "pep257"
 [tool.pytest.ini_options]
 addopts = "--doctest-modules"
 doctest_optionflags = "NORMALIZE_WHITESPACE"
-testpaths = "python"
+testpaths = ["python", "docs/examples/user_documentation.rst"]
 
