### 2025-05-23 994bde1
test: add json string test
