### 2025-05-23 e7610a0
feat: new sub-module to manipulate SF-MOCs
