+++ src/sfmoc.rs
@@ -1,12 +1,67 @@
 //! Methods releted to SF-MOC
 
-use numpy::{PyArrayMethods, PyReadonlyArrayDyn};
+use numpy::{PyArray1, PyArrayMethods, PyReadonlyArrayDyn};
 use pyo3::{
   exceptions::{PyIOError, PyValueError},
+  prelude::*,
   pyfunction, PyResult,
 };
 
-use moc::storage::u64idx::U64MocStore;
+use moc::{qty::Frequency, storage::u64idx::U64MocStore};
+
+/// Computes the depth of a Space-Frequency coverage
+///
+/// # Arguments
+///
+/// * ``index`` - The index of the Space-Frequency coverage.
+///
+/// # Infos
+///
+/// If the Space-Frequency coverage is empty, the returned
+/// depth is `(0, 0)`.
+#[pyfunction]
+pub fn coverage_sf_depth(index: usize) -> PyResult<(u8, u8)> {
+  U64MocStore::get_global_store()
+    .get_sfmoc_depths(index)
+    .map_err(PyIOError::new_err)
+}
+
+#[pyfunction]
+pub fn new_empty_sfmoc(depth_frequency: u8, depth_space: u8) -> PyResult<usize> {
+  U64MocStore::get_global_store()
+    .new_empty_sfmoc(depth_frequency, depth_space)
+    .map_err(PyIOError::new_err)
+}
+
+/// Returns the minimum frequency value of the Space-Frequency coverage
+///
+/// # Arguments
+///
+/// * ``index`` - The index of the Space-Frequency coverage.
+///
+#[pyfunction]
+pub fn coverage_sf_min_freq(index: usize) -> PyResult<f64> {
+  U64MocStore::get_global_store()
+    .get_1st_axis_min(index)
+    .and_then(|opt| opt.ok_or_else(|| String::from("Empty SF-MOC")))
+    .map(Frequency::<u64>::hash2freq)
+    .map_err(PyValueError::new_err)
+}
+
+/// Returns the maximum frequency value of the Space-Frequency coverage
+///
+/// # Arguments
+///
+/// * ``index`` - The index of the Space-Frequency coverage.
+///
+#[pyfunction]
+pub fn coverage_sf_max_freq(index: usize) -> PyResult<f64> {
+  U64MocStore::get_global_store()
+    .get_1st_axis_max(index)
+    .and_then(|opt| opt.ok_or_else(|| String::from("Empty SF-MOC")))
+    .map(Frequency::<u64>::hash2freq)
+    .map_err(PyValueError::new_err)
+}
 
 /// Deserialize a Frequency-Space coverage from a FITS file.
 ///
@@ -191,6 +246,67 @@ pub fn from_freq_ranges_lonlat(
     .map_err(PyValueError::new_err)
 }
 
+/// Create a Space-Frequency coverage from a list of
+/// (frequency_range, longitude, latitude, radius) tuples.
+///
+/// # Arguments
+///
+/// * ``frequencies_min`` - Minimum value for the frequency, in Hz
+/// * ``frequencies_max`` - Maximum value for the frequency, in Hz
+/// * ``d1`` - The depth along the frequency axis.
+/// * ``spatial_coverages`` - List of spatial coverages.
+///
+#[pyfunction]
+pub fn from_frequency_ranges_spatial_coverages(
+  frequencies_min: PyReadonlyArrayDyn<f64>,
+  frequencies_max: PyReadonlyArrayDyn<f64>,
+  d1: u8,
+  spatial_coverages: PyReadonlyArrayDyn<usize>,
+) -> PyResult<usize> {
[…]
