+++ src/lib.rs
@@ -1827,7 +1827,7 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
       .map_err(PyIOError::new_err)
       .and_then(|vec_bool| {
         Array::from_shape_vec(time_shape, vec_bool)
-          .map(|a| a.into_pyarray_bound(py))
+          .map(|a| a.into_pyarray(py))
           .map_err(|e| PyValueError::new_err(e.to_string()))
       })
   }
@@ -1859,7 +1859,7 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
     let it = it_time.zip(it_lon.zip(it_lat));
     U64MocStore::get_global_store()
       .filter_timepos(index, it, |b| b) // in numpy, the mask is reversed (true means do not select)
-      .map(|vec_bool| PyArray1::<bool>::from_vec_bound(py, vec_bool).to_owned())
+      .map(|vec_bool| PyArray1::<bool>::from_vec(py, vec_bool).to_owned())
       .map_err(PyIOError::new_err)
   }
 
@@ -1896,7 +1896,7 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
       .map_err(PyIOError::new_err)
       .and_then(|vec_bool| {
         Array::from_shape_vec(lon_shape, vec_bool)
-          .map(|a| a.into_pyarray_bound(py))
+          .map(|a| a.into_pyarray(py))
           .map_err(|e| PyValueError::new_err(e.to_string()))
       })
   }
@@ -1921,7 +1921,7 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
       .map_err(PyIOError::new_err)
       .and_then(|vec_bool| {
         Array::from_shape_vec(time_shape, vec_bool)
-          .map(|a| a.into_pyarray_bound(py))
+          .map(|a| a.into_pyarray(py))
           .map_err(|e| PyValueError::new_err(e.to_string()))
       })
   }
@@ -1946,7 +1946,7 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
       .map_err(PyIOError::new_err)
       .and_then(|vec_bool| {
         Array::from_shape_vec(time_shape, vec_bool)
-          .map(|a| a.into_pyarray_bound(py))
+          .map(|a| a.into_pyarray(py))
           .map_err(|e| PyValueError::new_err(e.to_string()))
       })
   }
@@ -1971,7 +1971,7 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
       .map_err(PyIOError::new_err)
       .and_then(|vec_bool| {
         Array::from_shape_vec(freq_shape, vec_bool)
-          .map(|a| a.into_pyarray_bound(py))
+          .map(|a| a.into_pyarray(py))
           .map_err(|e| PyValueError::new_err(e.to_string()))
       })
   }
@@ -1986,7 +1986,7 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
   fn to_fits_raw<'a>(py: Python<'a>, index: usize, pre_v2: bool) -> PyResult<Bound<'a, PyBytes>> {
     U64MocStore::get_global_store()
       .to_fits_buff(index, Some(pre_v2))
-      .map(move |b| PyBytes::new_bound(py, &b))
+      .map(move |b| PyBytes::new(py, &b))
       .map_err(PyIOError::new_err)
   }
 
@@ -2391,7 +2391,7 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
       }
     }
[…]
