+++ Cargo.toml
@@ -37,7 +37,7 @@ rayon = "1.10"
 num_threads = "0.1"
 
 [dependencies.numpy]
-version = "0.22"
+version = "0.25"
 
 [dependencies.ndarray]
 version = "0.16"
@@ -46,7 +46,7 @@ default-features = false # do not include the default features, and optionally
 features = ["rayon"]
 
 [dependencies.pyo3]
-version = "0.22"
+version = "0.25"
 features = ["extension-module"]
 
 [profile.release]
