+++ Cargo.toml
@@ -28,9 +28,9 @@ bench = true
 crate-type = ["cdylib"]
 
 [dependencies]
-# moc = { version = "0.17", features = ["storage"] }
+moc = { version = "0.18", features = ["storage"] }
 # moc = { git = 'https://github.com/cds-astro/cds-moc-rust', rev = '20a61a23cc38c7e31e17ab229faae0f6eb0d3a13', features = ["storage"] }
-moc = { git = 'https://github.com/cds-astro/cds-moc-rust', branch = 'main', features = ["storage"] }
+# moc = { git = 'https://github.com/cds-astro/cds-moc-rust', branch = 'main', features = ["storage"] }
 healpix = { package = "cdshealpix", version = "0.7" }
 # healpix = { package = "cdshealpix", git = 'https://github.com/cds-astro/cds-healpix-rust', branch = 'master' }
 rayon = "1.10"
