### 2025-05-26 f6b4e6f
feat: warn when degrade_to_order is called with more precise order [#188]
