+++ python/mocpy/fmoc/fmoc.py
@@ -138,6 +138,11 @@ class FrequencyMOC(AbstractMOC):
         >>> fmoc.degrade_to_order(10)
         10/752
         """
+        if new_order >= self.max_order:
+            warnings.warn(
+                "The new order is more precise than the current order, nothing done.",
+                stacklevel=2,
+            )
         index = mocpy.degrade(self.store_index, new_order)
         return FrequencyMOC(index)
 
