+++ python/mocpy/moc/moc.py
@@ -272,6 +272,11 @@ class MOC(AbstractMOC):
         `~mocpy.MOC`
             The degraded MOC.
         """
+        if new_order >= self.max_order:
+            warnings.warn(
+                "The new order is more precise than the current order, nothing done.",
+                stacklevel=2,
+            )
         index = mocpy.degrade(self.store_index, new_order)
         return MOC(index)
 
