### 2025-05-26 a9415e8
feat: add generic 'to_order'
