+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "MOCPy"
-version = "0.17.1"
+version = "0.18.0"
 authors = [
   "Matthieu Baumann <matthieu.baumann@astro.unistra.fr>",
   "Thomas Boch <thomas.boch@astro.unistra.fr>",
