### 2025-05-26 c9e9936
fix: notebooks
