+++ _sources/examples/user_documentation.rst.txt
@@ -207,3 +207,170 @@ Gallery of notebooks examples using FMOCs
 .. nbgallery::
     ../_collections/notebooks/First_Steps_with_FMOCs
 
+********************************
+SFMOC: Space-Frequency coverages
+********************************
+
+The `SFMOC` class allows manipulation and generation of
+**Space-Frequency Multi-Order Coverages** (SF-MOCs).
+
+Creating SF-MOC instances
+=========================
+
+New empty SF-MOC `~mocpy.SFMOC.new_empty`:
+
+    >>> from mocpy import SFMOC
+    >>> SFMOC.new_empty(max_order_frequency=40, max_order_space=12)
+    f40/ s12/
+
+Loading an SF-MOC from a FITS, JSON or ASCII file `~mocpy.SFMOC.load`:
+
+    >>> # we generate a fake ASCII file
+    >>> from tempfile import NamedTemporaryFile
+    >>> with NamedTemporaryFile(delete_on_close=False) as tf:
+    ...     tf.write(b"f20/0-100 s12/0-100")
+    ...     tf.close()
+    ...     # and we read it
+    ...     SFMOC.load(tf.name, format="ascii")
+    19
+    f14/0
+    15/2
+    18/24
+    20/100
+    s9/0
+    10/4-5
+    11/24
+    12/100
+    f20/ s12/
+
+This illustrated the load method that we use when we have a MOC in a file, but for
+ASCII strings, one can also directly use `~mocpy.SFMOC.from_string`:
+
+    >>> SFMOC.from_string("f15/2 s12/1")
+    f15/2
+    s12/1
+    f15/ s12/
+
+Creating SF-MOCs from physical parameters:
+
+1. From frequency-position "points"
+-----------------------------------
+
+In `~mocpy.SFMOC.from_frequencies_and_positions`, we give a set of frequency-position
+"points" and the SF-MOC will be built from the cells containing these punctual data at
+the maximum requested order. Use the utility functions
+`~mocpy.MOC.spatial_resolution_to_order` to chose the space resolution and
+`~mocpy.FrequencyMOC.relative_precision_to_order` to chose the frequency resolution.
+
+    >>> import astropy.units as u
+    >>> from mocpy import SFMOC
+    >>> SFMOC.from_frequencies_and_positions(
+    ...     [0.01, 0.1, 1, 10, 100] * u.Hz,
+    ...     lon=[0, 1, 2, 3, 4] * u.deg, lat=[0, 1, 2, 3, 4] * u.deg,
+    ...     max_order_frequency=25, max_order_space=8)
+    f25/22879928
+    s8/311296
+    f25/23750246
+    s8/311316
+    f25/24641536
+    s8/311378
+    f25/25493504
+    s8/311558
+    f25/26361856
+    s8/311624
+    f25/ s8/
+
+2. From frequency ranges associated to punctual positions
+---------------------------------------------------------
+
+With `mocpy.SFMOC.from_frequency_ranges_and_positions` we do not get a single cell in
+frequency, bu the cells covering a given frequency band/range:
+
+    >>> import astropy.units as u
+    >>> from mocpy import SFMOC
+    >>> SFMOC.from_frequency_ranges_and_positions(
+    ...     frequencies_min=[0.01, 0.1, 1, 10, 100] * u.Hz,
+    ...     frequencies_max=[0.1, 1, 10, 100, 1000] * u.Hz,
+    ...     lon=[0, 1, 2, 3, 4] * u.deg, lat=[0, 1, 2, 3, 4] * u.deg,
[…]
