+++ _sources/stubs/mocpy.SFMOC.rst.txt
@@ -1,0 +1,8 @@
+mocpy.SFMOC
+=====================
+
+.. currentmodule:: mocpy
+
+.. autoclass:: SFMOC
+    :members:
+    :inherited-members:
