### 2025-07-30 #238 @lwasser
Hey everyone! 👋🏻 i'm checking in on packages that need to move forward. it looks like this is an astropy package but also it needs pre-review checks. we will followup on next steps  soon.
