### 2025-08-04 a1622a6
Fix formatting on `install.html` page
