### 2025-08-04 631f681
Fix formatting on `install.html` page
