### 2025-08-04 a7fc24a
Update intersphinx links
A sphinx build surfaces several warnings on updated intersphinx locations. While nothing is currently broken, it's probably good to fix before it does.

A "default" intersphinx is imported from `sphinx_astropy` which contains links for python, numpy, matplotlib and astropy itself. Using that instead of hard-coding here shifts the burden to maintain links to the astropy team and avoids duplicated work. Packages not in that list are added to to dict here.
