### 2025-08-04 b63da35
Explain optional dependencies
While mocpy itself will work, I think it's a common use case to run the examples in the notebooks after installing mocpay. Thus, it's probably helpful to explain how to install the extra dependencies for the notebooks in the install directions in the docs.
