### 2025-08-04 #238 @hamogu
## Editor in Chief checks (done by @hamogu instead of the EIC)

---
---

Hi,
thanks to much for submitting the package and the apologize that it's taken us so long to get to it. PyOpenSci is *too* successful, and we're getting more submission right now than we can handle simultaneously, so it's been a bit of a delay. I've looked at your package and it's in pretty good shape - not surprisingly since it's been around for a while and has been astropy-affilated for a while, too. However, for PyOPenSci, we ask you to add a few more "housekeeping" files that don't have much to do with code, but with the community working on it and hopefully helping you to maintain it, e.g. a Code of conduct (see list above where tick marks are missing). Let me know know if you need any pointers with suggestions or examples!

When I looked at the package docs, I also noticed a few small things that could be fixed - I've opened issues and PRs for you since that's probably easier to keep track for you than me just writing it down here.

Please let me know when you've added the files so I can tick off the boxes below and we'll get busy looking for reviewers. (Though I already want to caution that August is a time when many potential reviewers are on vacation and we might have to wait for people to get back to work before we find reviewers.)
