### 2025-08-21 32dd98b
Doc update
* Rerun the notebooks to execute all cells. Still the display is broken for output cells that renders interactive display (ipyaladin and slider widget)
* Fix markdown on old notebooks
* Add toml as a dep for compiling the docs
