### 2025-08-21 db07330
:doc: fix critical warnings
