### 2025-08-21 35fbb6c
add astropy_healpix for generating the docs as well as it needs the example compute_borders which internally still uses astropy_healpix and not cdshealpix
