### 2025-08-21 b7a9a82
make this doctest pass, requesting the tableb1 return a 500 http error. Did the API of https://cdsarc.cds.unistra.fr/viz-bin/moc changed or did we get blacklisted somehow ?
