### 2025-08-21 26d82d3
change back to https://networkx.org/documentation/stable the networkx doc links
