### 2025-09-21 #238 @hamogu
@ManonMarchand : I see that you've addressed my issues, that looks good thank you!

However, for PyOPenSci, we ask you to add a few more "housekeeping" files that don't have much to do with code, but with the community working on it and hopefully helping you to maintain it, e.g. a Code of conduct (see list above where tick marks are missing). Let me know know if you need any pointers with suggestions or examples!
