### 2025-10-07 #238 @ManonMarchand
Hello, sorry for the delayed response.

Quick questions: Why is it imposed to have a code of conduct? We never needed one. Can we write our own with very minimal instructions?
