### 2025-10-07 #238 @hamogu
Of course you can write your own Code of Conduct. PyOpenSci is about fostering collaboration and we hope (and implicitly assume) that you will accept community contributions to your codebase, if people open PRs. As such, it's just good practice to write down a few words about how you expect contributors to behave, just to set expectations and to have something to point to if conflict arises. Most codes of conducts are a little more elaborate versions of "Be nice to each other and email XXX if you have a problem and if you Arne't nice we might ban you".
In my experience, most projects never need a code of conduct, but for a small fraction of projects conflicts arise at some point, people start yelling and insulting each other, etc. Hopefully, that never happens in MOCpy, but it it does, it can help to have a written down Code of Conduct that you can point to and say "These are the rules and this is what we do when you break them".  Think of it a bit like an insurance policy: You hope you never need it, but if conflict ever happens it might be helpful to have it.
