### 2026-02-19 #238 @pllim
As Brigitta pointed out on Astropy Slack, mocpy is already Astropy Affiliated before Astropy's partnership with pyOpenSci. So from Astropy's POV, this is a re-review. Also see:

* https://github.com/astropy/astropy.github.com/pull/319

I do not think we will sync the old record with the new info here (e.g., maintainers have changed) because https://www.astropy.org/affiliated/index.html#affiliated-package-list-pre-ape22 is frozen, right, @hamogu / @dhomeier ?
