### 2026-02-19 #238 @dhomeier
> I do not think we will sync the old record with the new info here (e.g., maintainers have changed) because https://www.astropy.org/affiliated/index.html#affiliated-package-list-pre-ape22 is frozen, right, [@hamogu](https://github.com/hamogu) / [@dhomeier](https://github.com/dhomeier) ?

Once accepted here, the pre-APE22 entry should probably be removed, since APE22 states that those badges should be considered outdated at a certain point and in any case would no longer reflect this package's status.
In principle the re-review might even be limited to the [health checks](https://github.com/astropy/astropy-APEs/blob/main/APE22.rst#re-review-of-packages) sketched in APE22, but as these had not been as clearly defined, and the original Astropy review criteria are not identical to the pyOpenSci ones, this will probably still become at least formally a full review.
