### 2026-02-20 #238 @pllim
Re: https://github.com/pyOpenSci/software-submission/issues/238#issuecomment-3930583747

Removing entry from that listing would be controversial. I don't think we ever hashed out all the details for a re-review back in https://github.com/astropy/astropy-APEs/blob/main/APE22.rst . I think we need to discuss that more back over on Astropy side, not here. Thanks!
