### 2026-02-20 #238 @hamogu
@pllim When packages get re-reviewed and listed as astropy-affiliated through the PyopenSci we can remove them from the pre-APE 22 list so that they are not listed twice. That [what we did with Stingray and Zodipy](https://github.com/astropy/astropy.github.com/pull/675) after they got accepted by PyOpenSci: We asked the package maintainers for permission and they agreed that it makes more sense otherwise have them listed only once on https://www.astropy.org/affiliated/index.html and so we removed them from the "old" listing.

However, that has no bearing on this review, it's a step we can discuss once this review here is done.
