### 2026-03-02 #238 @crhea93
## Editor response to review:

---

## Editor comments
:wave: Hi @TOCode11 and @ChristianThibeault! Thank you for volunteering to review for pyOpenSci!

- [ ] reviewer 3 (if applicable)

Please get in touch with any questions or concerns! Your review is due: <Insert deadline DATE HERE>
---

Reviewers: @TOCode11 @ChristianThibeault
Due date: April 15, 2026
