### 2026-04-16 #238 @TOCode11
- [X] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [X] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [See review comment 1)] **Examples** for all user-facing functions.
- [X] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

- [See review comment 2)] Badges for:
    - [See review comment 2)] Python versions supported,

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

- [See review comment 3)] Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the README.md file.
    - [See review comment 3)] Brief demonstration of package usage (as it makes sense - links to vignettes could also suffice here if package description is clear)

The package structure should follow the general community best practices. In general, please consider whether:

- [N/A] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](http://joss.theoj.org/about#submission_requirements).

- [N/A] **A short summary** describing the high-level functionality of the software
- [N/A] **Authors:** A list of authors with their affiliations
- [N/A] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
- [N/A] **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

---

0) Great work on the package. I enjoyed learning to use it. This is great work. My review is pretty modest as I don't have much experience with python development.

1) In docs/examples or in notebooks, I recommend adding a typical example for the use of class SFMOC, for completeness (unless I missed an example of its use?).

2) README.md doesn't currently have a badge for the python versions supported. I first tested version 0.17.1 and I had some python version dependencies issues. I suggest adding a badge for it or any type of indication of the python version supported.

3) README.md could contain a single vignette explaining basic usage of the package. An alternative would be to point to the notebooks or docs/examples with a clear indication of what to use as an onboarding for new users.

4) I had an issue with running docs/examples/plt_TMOC_SDSS_r.py. I opened an issue to describe it.

5) For JOSS submission, since I wasn't sure if it was submitted, I put N/A. If it's to be submitted to JOSS, then I believe the file paper.md is missing.
