### 2026-05-03 #238 @ChristianThibeault
- [X] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
  > **Reviewer:** README describe the IVOA MOC standard and what the package does. The target audience is implicit (astronomers, IVOA tooling) but understandable from context.
  > **Reviewer:** User install (`pip install mocpy`) is documented and works well. The development install path documented in the README did not work for me on my current Rust setup (See Review comments below).
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
  > **Reviewer:** Notebooks (14 of them) and `docs/examples/user_documentation.rst` cover most of the package.
  > **Reviewer:** All public methods across the five MOC classes have a docstring.
  > **Reviewer:** Less than half of public methods include a runnable `Examples` block. For example, no examples on `TimeMOC.from_time_ranges` or `STMOC.query_by_time`, two methods users hit early.
  > **Reviewer:** `CONTRIBUTING.md` is short and just redirects to the full guide at `docs/contribute.html`, which is fine.
- [ ] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.
  > **Reviewer:** `pyproject.toml` has URLs (Documentation, Repository, Issues, Changelog) and classifiers, but no `[project].authors` block. The four maintainers' names and emails appear in `Cargo.toml`. `CONTRIBUTING.md` tell users to find the emails *"in the `pyproject.toml` file"*, which is not the case

      > **Reviewer:** CI badge present (linking to `test.yml`). Coverage badge missing , coverage seems to be computed in CI but not shown anywhere in badges.
      > **Reviewer:** No Python-version badge in the README.

  > **Reviewer:** README cover stable (pip + conda), unreleased (`pip install git+https://...`), and development install. Although. Development install was broken for me on the Rust side (see below Review Comments - Suggested changes).
      > **Reviewer:** README contains a screenshot of a rendered MOC but no inline code example, no link to a vignette or relevant notebook files.
  > **Reviewer:** Doc links to https://cds-astro.github.io/mocpy/. (Great)

The package structure should follow the general community best practices. In general, please consider whether:

  > **Reviewer:** End-user install via `pip install mocpy` from PyPI works on my macOS / Python 3.11. The contributor install (`maturin develop --release`) failed for me on a clean Rust 1.85 install (See Review Comments).
  > **Reviewer:** No performance claims (no benchmarks, no timings, no "fast" wording) in the README or `docs/`. Nothing to verify.
    > **Reviewer:** 168 of 171 tests pass when run as `python -m pytest --pyargs mocpy.tests` from the repo root. Two failures (`test_mpl_fill`, `test_mpl_border`) seems to be caused by a matplotlib 3.10 version issue.
    > **Reviewer:** ~2,470 lines across 8 test files cover the public API of all five MOC classes
  > **Reviewer:** The test matrix is wide and clean. Coverage is computed but never shown as a badge or used as a threshold.
      > **Reviewer:** `pyproject.toml` declares `requires-python = ">=3.9"` and the classifiers list 3.9-3.14. Python 3.9 reached end-of-life last year (by the way: Python 3.10 will reach EOL on 2026-10-31).

> **Reviewer:** Not applicable. The package does not contain a `paper.md` and JOSS submission was not indicated.

  > **Reviewer:** Pending, approval conditional on the changes listed Review Comments section.

Estimated hours spent reviewing: ~10 hours.

---

##### Summary

MOCpy is a Python+Rust package that implements the IVOA MOC 2.0 standard, plus extensions for frequency and space-frequency coverages. Well maintained and have been used in production for a while for real astrophysics research.

Installation was straightforward. `pip install mocpy` from is clean. `import mocpy` succeeds. The main features I tested all work: cone and polygon MOCs, set operations, FITS I/O, time/space-time/frequency MOCs. The test suite passes 168 of 171 tests. Documentation is great: a Sphinx site, 14 notebooks, and 100% of public methods have a docstring. CI is wide: Linux, macOS, and Windows across Python 3.9 to 3.14, plus extra jobs for coverage docs, Pyodide, and codemeta (https://github.com/cds-astro/mocpy/blob/master/.github/workflows/test.yml). Deprecations are handled, with `DeprecationWarning` and tests that check the warning is raised.

That said, I found some things to fix (which I annotated under the unchecked bullet points above). None of them affect end-users who install from PyPI today, but some breaks a pyOpenSci checklist item.

##### Suggested changes

- **R1 - Drop Python 3.9:**`pyproject.toml` says `requires-python = ">=3.9"` and the classifiers list 3.9. But Python 3.9 reached end of life on last year. pyOpenSci does not allow EOL versions. Bump the minimum to `>=3.10`. Note that 3.10 itself ends this year, so going to `>=3.11` could make sense too.
- **R2 - Add `[project].authors` to `pyproject.toml` and fix `CONTRIBUTING.md:`** The four maintainers' names and emails are only in `Cargo.toml`. The `[project]` block has no `authors` field, so PyPI shows no author at all. Also, `CONTRIBUTING.md` lines 11-13 tell people to find the emails in `pyproject.toml`, but they are not there. You can copy the four entries from `Cargo.toml` into a `[project].authors` block.
- **R3 - Fix the doctest in `docs/examples/user_documentation.rst`:** Line 231 uses `NamedTemporaryFile(delete_on_close=False)`. That keyword was added in Python 3.12. The package supports 3.9+, so the doctest fails on 3.9, 3.10, and 3.11. Replace it with `delete=False`, which works on every supported version. This file is the canonical "quickstart that runs locally", so a broken doctest here is not a small detail. Note that R1's recommended bump (`>=3.10`, or `>=3.11`) does not fix this on its own, only a bump to `>=3.12` would.
- **R4 - Make the development install easier for non-Rust contributors:** I am not very familiar with Rust and its package installations, I had trouble following the "Development installation" steps in the README on a clean machine. I had errors of the form:

```
error[E0599]: no function or associated item named `from_utf8_unchecked` found for type `str` in the current scope
   --> .../cdshealpix-0.9.1/src/nested/sort/scindex.rs:354:14
    |
354 |         str::from_utf8_unchecked(&indexed_file_mdr5)
    |              ^^^^^^^^^^^^^^^^^^^ function or associated item not found in `str`
```
	End users are not affected (PyPI and conda-forge ship prebuilt wheels), so this is purely for contributors. It would help to make the development installation more approachable for people who are not Rust developers.
- **R5 - Expand Docstring Examples:** Currently, less than half of public methods include runnable `Examples` blocks. While the provided notebooks are excellent, many users rely on IDE tooltips (via Language Servers) and the API reference page. Prioritize adding examples to high-traffic methods such as `TimeMOC.from_time_ranges` and `STMOC.query_by_time` to make the developer experience more seamless.

##### Minor suggestions

- **Add a code example to the README.** It currently has a screenshot of a rendered MOC and zero lines of code. A 5-8 line block (cone MOC, sky fraction) right after the description would help a lot for first readers/users.
- in README: Confusing sentence and typo: *"We strongly recommend to work in an environnement"*
- **Add type annotations slowly.** Across the six main modules, about 1% are fully annotated. Suggest to start with the most-used public methods.
- **Fix the dependency list in `docs/install.rst`.** It says: *"MOCPy relies on the following mandatory external dependencies: cdshealpix, astropy, matplotlib, networkx"*. But `pyproject.toml` only marks `astropy` and `numpy` as mandatory; `cdshealpix`, `matplotlib`, and `networkx` are in the `plots` extra and are only needed if the user calls the plotting methods (`MOC.fill`, `MOC.border`, `MOC.get_boundaries`). The pyproject is the right shape, but the install page contradicts it and will lead users to think `pip install mocpy` is not enough for basic MOC operations.
- **Track the 9 TODOs.** `abstract_moc.py` , `moc.py`, `boundaries.py,` `test_moc.py:`. They are all transparent (good), but I figure I would mention it.
