### 2026-08-19 #238 @ManonMarchand
The team and I would like to thank the reviewers for the time spent reading MOCPy and for all the constructive comments and for the spotted issues! ✨

We tried to apply everything and here is a point per point answer:

## Review 1

> In docs/examples or in notebooks, I recommend adding a typical example for the use of class SFMOC, for completeness (unless I missed an example of its use?).

It's true that we don't have a concrete example yet for SFMOCs.
It's because Frequency MOCs (and thus Space-Frequency MOCs too) are still a draft standard.
We're waiting for the first data centers to publish SFMOCs and for the first papers to start using them to see what could be interesting in a notebook.
It was strongly requested from solar system meteorologists but that's not our lab's specialty, we were happy to add the support for the draft standard, and we provided a light description of the API. For now it looks like the researchers are playing with MOCPy but it's still ongoing.

> README.md doesn't currently have a badge for the python versions supported. I first tested version 0.17.1 and I had some python version dependencies issues. I suggest adding a badge for it or any type of indication of the python version supported.

Do you remember which version of python was not working for you? In theory we should support 3.9 to 3.14 (without free-threading) for now.

The badge was added in https://github.com/cds-astro/mocpy/pull/221

> README.md could contain a single vignette explaining basic usage of the package. An alternative would be to point to the notebooks or docs/examples with a clear indication of what to use as an onboarding for new users.

Thanks, done in https://github.com/cds-astro/mocpy/pull/228

> I had an issue with running docs/examples/plt_TMOC_SDSS_r.py. I opened an issue to describe it.

Thank you for noticing the issue. The root of the problem was that we were trying to calculate the time corresponding to the mouse position event when the mouse was hovering outside of the plot area. This is fixed in https://github.com/cds-astro/mocpy/pull/220

## Review 2

> README contains a screenshot of a rendered MOC but no inline code example, no link to a vignette or relevant notebook files.

Thank you for the suggestion, done in https://github.com/cds-astro/mocpy/pull/228

> 168 of 171 tests pass when run as python -m pytest --pyargs mocpy.tests from the repo root. Two failures (test_mpl_fill, test_mpl_border) seems to be caused by a matplotlib 3.10 version issue.

I cannot reproduce these failures with matplotlib 3.10.

I get some somewhat cryptic warnings `.conda/envs/py310/lib/python3.10/site-packages/matplotlib/_fontconfig_pattern.py:64: PyparsingDeprecationWarning: 'oneOf' deprecated - use 'one_of'` and `.conda/envs/py310/lib/python3.10/site-packages/matplotlib/_fontconfig_pattern.py:85: PyparsingDeprecationWarning: 'parseString' deprecated - use 'parse_string'
    parse = parser.parseString(pattern)` but they disappear with more recent versions of matplotlib.

> Coverage is computed but never shown as a badge or used as a threshold.

Yes. We use it internally and the implicit rule is that new PRs should test anything they touched. It works pretty well. Do you think that a coverage badge or action is mandatory?

>  Drop Python 3.9:pyproject.toml says requires-python = ">=3.9" and the classifiers list 3.9. But Python 3.9 reached end of life on last year. pyOpenSci does not allow EOL versions. Bump the minimum to >=3.10. Note that 3.10 itself ends this year, so going to >=3.11 could make sense too.

We agree that our approach is different from the rest of the python ecosystem. It's a
compromise. Our users often work with long running projects (telescope pipelines for
example) and they cannot easily update their python version but still want the bug fixes
in mocpy.
We don't have the maintenance power to backport all
bugfixes in the older versions so we decided to always support all alive python version
& the latest EOL. It seems to work for both the maintenance team and our users.

> Add [project].authors to pyproject.toml and fix CONTRIBUTING.md: The four maintainers' names and emails are only in Cargo.toml. The [project] block has no authors field, so PyPI shows no author at all. Also, CONTRIBUTING.md lines 11-13 tell people to find the emails in pyproject.toml, but they are not there. You can copy the four entries from Cargo.toml into a [project].authors block.

Thank you! Done in https://github.com/cds-astro/mocpy/pull/229

> Fix the doctest in docs/examples/user_documentation.rst: Line 231 uses NamedTemporaryFile(delete_on_close=False). That keyword was added in Python 3.12. The package supports 3.9+, so the doctest fails on 3.9, 3.10, and 3.11. Replace it with delete=False, which works on every supported version. This file is the canonical "quickstart that runs locally", so a broken doctest here is not a small detail. Note that R1's recommended bump (>=3.10, or >=3.11) does not fix this on its own, only a bump to >=3.12 would.

Thank you, we completely missed this!
The issue is fixed in https://github.com/cds-astro/mocpy/pull/230

> Make the development install easier for non-Rust contributors: I am not very familiar with Rust and its package installations, I had trouble following the "Development installation" steps in the README on a clean machine.

We now declare explicitly the minimal supported Rust version in the `Cargo.toml`.
We added instructions to check that the rust version is compatible in the README,
and we check with a new CI in `check-rust-versions.yml` that we indeed can
compile with the declared oldest supported rust version.
See https://github.com/cds-astro/mocpy/pull/218

Thank you for your remark on this!

> Expand Docstring Examples: Currently, less than half of public methods include runnable Examples blocks. While the provided notebooks are excellent, many users rely on IDE tooltips (via Language Servers) and the API reference page. Prioritize adding examples to high-traffic methods such as TimeMOC.from_time_ranges and STMOC.query_by_time to make the developer experience more seamless.

Pfiou that was quite a task! Done in https://github.com/cds-astro/mocpy/pull/234
Thanks for the incentive to do it ;)

> in README: Confusing sentence and typo: "We strongly recommend to work in an environnement"

Thank you! Fixed in https://github.com/cds-astro/mocpy/commit/08ef91460d56d3156807d042b11ecd5fea20727b

> Add type annotations slowly. Across the six main modules, about 1% are fully annotated. Suggest to start with the most-used public methods.

Yes, we'll add this in our radar, surely as a milestone for the version 1.0 release.

> Fix the dependency list in docs/install.rst. It says: "MOCPy relies on the following mandatory external dependencies: cdshealpix, astropy, matplotlib, networkx". But pyproject.toml only marks astropy and numpy as mandatory; cdshealpix, matplotlib, and networkx are in the plots extra and are only needed if the user calls the plotting methods (MOC.fill, MOC.border, MOC.get_boundaries). The pyproject is the right shape, but the install page contradicts it and will lead users to think pip install mocpy is not enough for basic MOC operations.

Thank you for spotting this. These dependencies used to be mandatory and we forgot to change
the documentation when this changed. Done in https://github.com/cds-astro/mocpy/pull/231

> Track the 9 TODOs. abstract_moc.py , moc.py, boundaries.py, test_moc.py:. They are all transparent (good), but I figure I would mention it.

Hum yes... Do you mean that we should open a github issue for each TODO? These are mostly internal notes between the maintainers and potential contributors
