### 2026-08-31 #238 @crhea93
@ChristianThibeault @TOCode11 Salut! Just checking to see if you two would mind looking into these responses :D
