### 2026-09-03 #238 @TOCode11
I used Python 3.14.2. @ManonMarchand
