### 2026-09-03 #238 @crhea93
@ManonMarchand : @ChristianThibeault is out of town until mid September. When he gets back he will get to the responses 😄
